# GH-3585 improve sorting (with CoreDatatype)