--- a/core/model-api/src/main/java/org/eclipse/rdf4j/model/ValueFactory.java
+++ b/core/model-api/src/main/java/org/eclipse/rdf4j/model/ValueFactory.java
@@ -101,6 +101,16 @@ public interface ValueFactory {
 	Literal createLiteral(String label, CoreDatatype datatype);
 
 	/**
+	 * Creates a new literal with the supplied label and datatype.
+	 *
+	 * @param label    The literal's label, must not be <var>null</var>.
+	 * @param datatype The literal's datatype. If it is null, the datatype
+	 *                 <a href="http://www.w3.org/2001/XMLSchema#string">{@code xsd:string}</a> will be assigned to this
+	 *                 literal.
+	 */
+	Literal createLiteral(String label, IRI datatype, CoreDatatype coreDatatype);
+
+	/**
 	 * Creates a new <var>xsd:boolean</var>-typed literal representing the specified value.
 	 *
 	 * @param value The value for the literal.