--- a/core/model/src/main/java/org/eclipse/rdf4j/model/impl/DecimalLiteral.java
+++ b/core/model/src/main/java/org/eclipse/rdf4j/model/impl/DecimalLiteral.java
@@ -29,7 +29,7 @@ public class DecimalLiteral extends SimpleLiteral {
 	 * Creates an xsd:decimal literal with the specified value.
 	 */
 	protected DecimalLiteral(BigDecimal value) {
-		this(value, XSD.DECIMAL);
+		this(value, CoreDatatype.XSD.DECIMAL);
 	}
 
 	/**