@hmottestad 2022-01-19
## Before
```
Benchmark                        Mode  Cnt     Score     Error  Units
SortBenchmark.sortByQuery        avgt   10  1589.908 ± 562.493  ms/op
SortBenchmark.sortDirectly       avgt   10   425.071 ±  11.140  ms/op
SortBenchmark.sortGetStatements  avgt   10   379.233 ±   4.398  ms/op
```

## After
```
Benchmark                        Mode  Cnt    Score    Error  Units
SortBenchmark.sortByQuery        avgt   10  555.001 ± 14.542  ms/op
SortBenchmark.sortDirectly       avgt   10  150.268 ±  3.928  ms/op
SortBenchmark.sortGetStatements  avgt   10  157.700 ±  1.197  ms/op
```